refactor(core): Split tiles.ts into category-based modules for LLM optimization
Split the 684-line tiles.ts file into focused modules:
- tiles/types.ts - LocalHex, TileDefinition, TileType interfaces
- tiles/hexPositions.ts - LOCAL_HEX constant
- tiles/starting.ts - Starting tile definitions (2 tiles)
- tiles/countrysideBase.ts - Base game countryside tiles (11 tiles)
- tiles/countrysideExpansion.ts - Lost Legion countryside tiles (3 tiles)
- tiles/countryside.ts - Re-exports all countryside tiles
- tiles/coreNonCity.ts - Base game non-city core tiles (4 tiles)
- tiles/coreCity.ts - Base game city core tiles (4 tiles)
- tiles/coreExpansion.ts - Lost Legion core tiles (3 tiles)
- tiles/core.ts - Re-exports all core tiles
- tiles/helpers.ts - placeTile, getTilesByType, helper functions
- tiles/index.ts - Aggregated exports and re-exports

All exports preserved, imports updated in consuming files.
